bun,app: upgrade next, pull userId from server
