+++ package.json
@@ -50,7 +50,7 @@
     "logrocket": "^5.0.0",
     "lucide-react": "^0.268.0",
     "luxon": "^3.4.0",
-    "next": "^13.5.2",
+    "next": "14.0.3",
     "next-themes": "^0.2.1",
     "openai": "^3.2.1",
     "pg": "^8.11.3",
